// src/payload/payload-types.ts
export type CartItems = {
  product?: string | Product;
  quantity?: number;
  id?: string;
}[];
export type FavouriteItems = {
  product?: string | Product;
  id?: string;
}[];
// ...
export interface User {
  id: string;
  name?: string;
  phone?: string;
  roles?: ('admin' | 'customer')[];
  purchases?: string[] | Product[];
  stripeCustomerID?: string;
  cart?: {
    items?: CartItems;
  };
  favourites?: {
    items?: FavouriteItems;
  };
  skipSync?: boolean;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string;
  resetPasswordExpiration?: string;
  salt?: string;
  hash?: string;
  loginAttempts?: number;
  lockUntil?: string;
  password: string;
}